requirements.txt
cryptography
pyperclip
argon2-cffi
flask
flask-jwt-extended